refactor: sns-login → sns-login-jwt 폴더 및 모듈 경로 변경
