+++ web/sns-login-jwt/backend/Dockerfile
