+++ web/sns-login-jwt/backend/main.go
@@ -3,13 +3,13 @@ package main
 import (
 	"log"
 
-	"github.com/kenshin579/tutorials-go/web/sns-login/backend/config"
-	"github.com/kenshin579/tutorials-go/web/sns-login/backend/handler"
-	customMiddleware "github.com/kenshin579/tutorials-go/web/sns-login/backend/middleware"
-	"github.com/kenshin579/tutorials-go/web/sns-login/backend/model"
-	"github.com/kenshin579/tutorials-go/web/sns-login/backend/provider"
-	"github.com/kenshin579/tutorials-go/web/sns-login/backend/repository"
-	"github.com/kenshin579/tutorials-go/web/sns-login/backend/service"
+	"github.com/kenshin579/tutorials-go/web/sns-login-jwt/backend/config"
+	"github.com/kenshin579/tutorials-go/web/sns-login-jwt/backend/handler"
+	customMiddleware "github.com/kenshin579/tutorials-go/web/sns-login-jwt/backend/middleware"
+	"github.com/kenshin579/tutorials-go/web/sns-login-jwt/backend/model"
+	"github.com/kenshin579/tutorials-go/web/sns-login-jwt/backend/provider"
+	"github.com/kenshin579/tutorials-go/web/sns-login-jwt/backend/repository"
+	"github.com/kenshin579/tutorials-go/web/sns-login-jwt/backend/service"
 	"github.com/labstack/echo/v4"
 	echoMiddleware "github.com/labstack/echo/v4/middleware"
 	"gorm.io/driver/sqlite"
