+++ web/sns-login-jwt/backend/middleware/auth_middleware.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 	"strings"
 
-	"github.com/kenshin579/tutorials-go/web/sns-login/backend/service"
+	"github.com/kenshin579/tutorials-go/web/sns-login-jwt/backend/service"
 	"github.com/labstack/echo/v4"
 )
