+++ web/sns-login-jwt/frontend/Dockerfile
