+++ web/sns-login-jwt/frontend/nginx.conf
