+++ web/sns-login-jwt/frontend/src/components/UserProfile.tsx
