+++ web/sns-login-jwt/frontend/src/types/auth.ts
