+++ web/sns-login-jwt/frontend/tsconfig.tsbuildinfo
